requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==45.0.3
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.1
pytest==8.4.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
